Refactor objects.rs  (#527)
* Remove objets.rs and add everything in mod.rs

* import only used structs

* fix merge

* cargo fmt

---------
