+++ src/business_logic/execution/gas_usage.rs
@@ -1,4 +1,4 @@
-use super::objects::L2toL1MessageInfo;
+use crate::business_logic::execution::L2toL1MessageInfo;
 use crate::definitions::constants::*;
 use crate::services::eth_definitions::eth_gas_constans::*;
 
@@ -130,10 +130,7 @@ pub fn get_event_emission_cost(topics: usize, l1_handler_payload_size: usize) ->
 mod test {
     use super::get_event_emission_cost;
     use super::*;
-    use crate::{
-        business_logic::execution::objects::{L2toL1MessageInfo, OrderedL2ToL1Message},
-        utils::Address,
-    };
+    use crate::{business_logic::execution::OrderedL2ToL1Message, utils::Address};
     use coverage_helper::test;
 
     #[test]
