+++ src/business_logic/transaction/objects/internal_declare.rs
@@ -1,8 +1,8 @@
 use crate::{
     business_logic::{
         execution::{
-            execution_entry_point::ExecutionEntryPoint,
-            objects::{CallInfo, TransactionExecutionContext, TransactionExecutionInfo},
+            execution_entry_point::ExecutionEntryPoint, CallInfo, TransactionExecutionContext,
+            TransactionExecutionInfo,
         },
         fact_state::state::ExecutionResourcesManager,
         state::state_api::{State, StateReader},
@@ -307,8 +307,7 @@ mod tests {
 
     use crate::{
         business_logic::{
-            execution::objects::{CallInfo, CallType, TransactionExecutionInfo},
-            fact_state::in_memory_state_reader::InMemoryStateReader,
+            execution::CallType, fact_state::in_memory_state_reader::InMemoryStateReader,
             state::cached_state::CachedState,
         },
         definitions::{
