+++ src/business_logic/transaction/objects/internal_invoke_function.rs
@@ -3,8 +3,8 @@ use std::collections::HashMap;
 use crate::{
     business_logic::{
         execution::{
-            execution_entry_point::ExecutionEntryPoint,
-            objects::{CallInfo, TransactionExecutionContext, TransactionExecutionInfo},
+            execution_entry_point::ExecutionEntryPoint, CallInfo, TransactionExecutionContext,
+            TransactionExecutionInfo,
         },
         fact_state::state::ExecutionResourcesManager,
         state::state_api::{State, StateReader},
