+++ tests/cairo_1_syscalls.rs
@@ -14,10 +14,8 @@ use num_traits::{Num, One, Zero};
 use starknet_rs::{
     business_logic::{
         execution::{
-            execution_entry_point::ExecutionEntryPoint,
-            objects::{
-                CallInfo, CallType, OrderedEvent, OrderedL2ToL1Message, TransactionExecutionContext,
-            },
+            execution_entry_point::ExecutionEntryPoint, CallInfo, CallType, OrderedEvent,
+            OrderedL2ToL1Message, TransactionExecutionContext,
         },
         fact_state::{
             in_memory_state_reader::InMemoryStateReader, state::ExecutionResourcesManager,
