+++ tests/complex_contracts/amm_contracts/amm.rs
@@ -5,7 +5,7 @@ use cairo_vm::vm::runners::cairo_runner::ExecutionResources;
 use num_traits::Zero;
 use starknet_rs::{
     business_logic::{
-        execution::objects::{CallInfo, CallType},
+        execution::{CallInfo, CallType},
         fact_state::{
             in_memory_state_reader::InMemoryStateReader, state::ExecutionResourcesManager,
         },
