+++ tests/complex_contracts/amm_contracts/amm_proxy.rs
@@ -4,7 +4,7 @@ use cairo_vm::vm::runners::cairo_runner::ExecutionResources;
 use starknet_crypto::FieldElement;
 use starknet_rs::{
     business_logic::{
-        execution::objects::{CallInfo, CallType},
+        execution::{CallInfo, CallType},
         fact_state::{
             in_memory_state_reader::InMemoryStateReader, state::ExecutionResourcesManager,
         },
