+++ tests/complex_contracts/utils.rs
@@ -6,8 +6,8 @@ use starknet_crypto::{pedersen_hash, FieldElement};
 use starknet_rs::{
     business_logic::{
         execution::{
-            execution_entry_point::ExecutionEntryPoint,
-            objects::{CallInfo, CallType, TransactionExecutionContext},
+            execution_entry_point::ExecutionEntryPoint, CallInfo, CallType,
+            TransactionExecutionContext,
         },
         fact_state::{
             in_memory_state_reader::InMemoryStateReader, state::ExecutionResourcesManager,
