+++ tests/internal_calls.rs
@@ -5,8 +5,7 @@ use num_traits::Zero;
 use starknet_rs::{
     business_logic::{
         execution::{
-            execution_entry_point::ExecutionEntryPoint,
-            objects::{CallType, TransactionExecutionContext},
+            execution_entry_point::ExecutionEntryPoint, CallType, TransactionExecutionContext,
         },
         fact_state::{
             in_memory_state_reader::InMemoryStateReader, state::ExecutionResourcesManager,
